Выносит функции по работе с комментариями в отдельный модуль
