+++ js/main.js
@@ -1,7 +1,7 @@
 import {createPhotosDescriptions} from './data.js';
 import {buildThumbnails} from './thumbnails-rendering.js';
 import {createFullPicture} from './full-size-picture.js';
-import {validateForm} from './form-validation.js'
+import {validateForm} from './form-validation.js';
 const photosDescriptions = createPhotosDescriptions(25);
 const photosContainer = document.querySelector('.pictures');
 const photosFragment = buildThumbnails(photosDescriptions);
